Refactoring and modular serpartion
Closes #5
