+++ index.html
@@ -6,7 +6,7 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/styles.css">
   <link href="https://fonts.googleapis.com/css?family=Open+Sans|Work+Sans:300" rel="stylesheet">
-  <script src="scripts/main.js" defer></script>
+  <script src="scripts/main.js" type= "module" defer></script>
   <title>Pomo timer</title>
 </head>
 <body>
